Replaced ssize_t with int (problems with this type with VC++ 6 and well... it does mot matter as the buffer size to do the copy is limited to 1 MB).
git-svn-id: https://vdsf.svn.sourceforge.net/svnroot/vdsf/trunk@732 53c26a9f-4021-0410-9f73-c4284c400f72

+++ src/Common/MemoryFile.c
@@ -194,7 +194,7 @@ int vdscCopyBackstore( vdscMemoryFile*   pMem,
    char bckName [PATH_MAX];
    size_t length = 1024*1024, leftToCopy;
    unsigned char * buffer = NULL;
-   ssize_t i, j, k;
+   int i, j, k;
    
    VDS_PRE_CONDITION( pMem    != NULL );
    VDS_INV_CONDITION( pMem->initialized == VDSC_MEMFILE_SIGNATURE );
